+++ packages/liveblocks-core/src/index.ts
@@ -37,7 +37,7 @@ export type {
   StringifyCommentBodyOptions,
 } from "./comments/comment-body";
 export {
-  getMentionedIdsFromCommentBody,
+  getMentionsFromCommentBody,
   html,
   htmlSafe,
   isCommentBodyLink,
@@ -334,6 +334,7 @@ export type {
 } from "./types/PlainLson";
 export type { User } from "./types/User";
 export { detectDupes };
+export type { MentionData, UserMentionData } from "./types/MentionData";
 
 /**
  * Helper type to help users adopt to Lson types from interface definitions.